Expand pyright to cover tests/ and reproducibility/
- Add __iter__ to NetworkXView so list() works on dataset views
- Cast np.quantile/mean/std to float in MetricInterval.from_samples
- Replace BinomConfidenceInterval namedtuple with typed class
- Fix Literal type mismatches for split/variant params at call sites
- Add type narrowing assertions in tests for Optional attributes
- Fix matplotlib private imports (use ticker/colors modules directly)
- Exclude third-party test implementations (ggm/gran) from pyright
- Fix pre-commit pyright hook to use pixi run

+++ polygraph/metrics/base/metric_interval.py
@@ -51,14 +51,16 @@ class MetricInterval:
 
         if coverage is not None:
             low, high = (
-                np.quantile(samples, (1 - coverage) / 2, axis=0),
-                np.quantile(samples, coverage + (1 - coverage) / 2, axis=0),
+                float(np.quantile(samples, (1 - coverage) / 2, axis=0)),
+                float(
+                    np.quantile(samples, coverage + (1 - coverage) / 2, axis=0)
+                ),
             )
         else:
             low, high = None, None
 
-        mean = np.mean(samples, axis=0)
-        std = np.std(samples, axis=0)
+        mean = float(np.mean(samples, axis=0))
+        std = float(np.std(samples, axis=0))
 
         return cls(mean=mean, std=std, low=low, high=high, coverage=coverage)
